# components/backend/src/main/resources/restaurant-api-doc.yaml
openapi: 3.0.3
info:
  title: Swagger Restaurant
  description: |-
    This is a sample Restaruant
  version: 1.0.1
...
servers:
  - url: http://localhost:8084
...
paths:
...
  /user/{userId}:
    parameters:
      - name: userId
        in: path
        description: The ID of the user
        required: true
        schema:
          type: string
          example: "ObjectId('6544cd596955fe0a1c04fba9')"
    get:
      tags:
        - user
      summary: Get user by ID
      operationId: getUserById
      security:
      - sessionId: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    put:
      tags:
        - user
      summary: Update user by ID
      operationId: updateUserById
      security:
      - sessionId: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
        '405':
          description: Validation exception
    delete:
      tags:
        - user
      summary: Delete user by ID
      operationId: deleteUserById
      security:
      - sessionId: []
      responses:
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
...
  /table-order:
    get:
      tags:
        - table-order
      summary: Query table orders
      description: Place a new order in the store
      operationId: queryTableOrders
      security:
      - sessionId: []
      parameters:
        - name: table-number
          in: header
          schema: { type: integer }
          required: false
        - name: is-current
          in: header
          schema: { type: boolean }
          required: false
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '405':
          description: Invalid input
  /table-order/{table-number}:
    get:
      tags:
        - table-order
      summary: Find the current order for the table
      description: Place a new order in the store
      operationId: findTableOrder
      security:
      - sessionId: []
      parameters:
        - name: table-number
          in: path
          schema: { type: integer }
          required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '405':
          description: Invalid input

    post:
      tags:
        - table-order
      summary: Place an order
      description: Place a new order in the store
      operationId: placeOrder
      security:
      - sessionId: []
      parameters:
        - name: table-number
          in: path
          schema: { type: integer }
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '405':
          description: Invalid input
...
components:
  schemas:
...
    Food:
      required:
        - name
        - price
      type: object
      properties:
        id:
          type: string
          example: 6544cd596955fe0a1c04fba9
        name:
          type: string
          example: Hamburger
        price:
          type: integer
          example: 400
        recipe:
          type: array
          items:
            type: object
            properties:
              ingerient:
                type: string
              unit:
                type: string
                enum:
                  - g
                  - dkg
              quantity:
                type: integer
    FoodOrderItem:
        required:
          - food
          - quantity
        type: object
        properties:
          food: 
            $ref: '#/components/schemas/Food'
          quantity:
            type: integer
            format: int32
            example: 1
    Drink:
      required:
        - name
        - price
      type: object
      properties:
        id:
          type: string
          example: 6544cd596955fe0a1c04fba9
        name:
          type: string
          example: Cola
        price:
          type: integer
          example: 400
    DrinkOrderItem:
        required:
          - drink
          - quantity
        type: object
        properties:
          drink: 
            $ref: '#/components/schemas/Drink'
          quantity:
            type: integer
            format: int32
            example: 1
    Order:
      required:
        - foods
        - drinks
      type: object
      properties:
        id:
          type: string
          example: 6544cd596955fe0a1c04fba9
        status:
          type: string
          enum: [placed, cooking, delivering, delivered]
        date:
          type: string
          format: date-time
        note:
          type: string
        foods:
          type: array
          items:
            $ref: '#/components/schemas/FoodOrderItem'
        drinks:
          type: array
          items:
            $ref: '#/components/schemas/DrinkOrderItem'
...
    User:
      required:
        - name
        - email
        - mobil
        - address
      type: object
      properties:
        id:
          type: string
          example: ObjectId('6544cd596955fe0a1c04fba9')
        name:
          type: string
          example: Gipsz Jakap
        email:
          type: string
          example: gipsz@jakab.com
        mobil:
          type: string
          example: +36709834234
        address:
          type: string
          example: Budapest Lakatos utca 6.
...
  securitySchemes:
    sessionId:
      type: apiKey
      in: header
      name: sessionId